# specs/001-ai-inventory-voice/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory Management System API
  description: |
    RESTful API for AI-powered inventory management with voice interaction capabilities.
    Multi-tenant, role-based access control, real-time updates.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token with tenant_id and user roles
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    Product:
      type: object
      required:
        - sku
        - name
        - unit_of_measure
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        sku:
          type: string
          maxLength: 100
        name:
          type: string
          maxLength: 255
        description:
          type: string
        category:
          type: string
          maxLength: 100
        unit_of_measure:
          type: string
          enum: [pieces, kg, liters, meters, boxes]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Warehouse:
      type: object
      required:
        - name
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postal_code:
          type: string
        capacity_total:
          type: number
          format: double
        capacity_unit:
          type: string
        is_active:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Inventory:
      type: object
      required:
        - product_id
        - warehouse_id
        - quantity
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        warehouse_id:
          type: string
          format: uuid
        quantity:
          type: number
          format: double
          minimum: 0
        reserved_quantity:
          type: number
          format: double
          minimum: 0
        minimum_stock:
          type: number
          format: double
        safety_stock:
          type: number
          format: double
        reorder_point:
          type: number
          format: double
        last_movement_at:
          type: string
          format: date-time
        last_updated_at:
          type: string
          format: date-time
        product:
          $ref: '#/components/schemas/Product'
        warehouse:
          $ref: '#/components/schemas/Warehouse'
    InventoryMovement:
      type: object
      required:
        - movement_type
        - product_id
        - quantity
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        movement_type:
          type: string
          enum: [inbound, outbound, transfer]
        product_id:
          type: string
          format: uuid
        source_warehouse_id:
          type: string
          format: uuid
          nullable: true
        destination_warehouse_id:
          type: string
          format: uuid
          nullable: true
        quantity:
          type: number
          format: double
          minimum: 0.001
        quantity_before:
          type: number
          format: double
        quantity_after:
          type: number
          format: double
        reference_number:
          type: string
          maxLength: 100
        notes:
          type: string
        performed_by:
          type: string
          format: uuid
        performed_at:
          type: string
          format: date-time
    MovementRequest:
      type: object
      required:
        - movement_type
        - product_id
        - quantity
      properties:
        movement_type:
          type: string
          enum: [inbound, outbound, transfer]
        product_id:
          type: string
          format: uuid
        source_warehouse_id:
          type: string
          format: uuid
          nullable: true
        destination_warehouse_id:
          type: string
          format: uuid
          nullable: true
        quantity:
          type: number
          format: double
          minimum: 0.001
        reference_number:
          type: string
        notes:
          type: string
...
paths:
  /products:
    get:
      tags: [Products]
      summary: List products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
            maximum: 1000
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
    post:
      tags: [Products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /products/{product_id}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Products]
      summary: Update product
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Products]
      summary: Delete product
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Product deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /inventory:
    get:
      tags: [Inventory]
      summary: List inventory
      parameters:
        - name: product_id
          in: query
          schema:
            type: string
            format: uuid
        - name: warehouse_id
          in: query
          schema:
            type: string
            format: uuid
        - name: low_stock
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of inventory records
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inventory'
                  total:
                    type: integer
  /inventory/movement:
    post:
      tags: [Inventory]
      summary: Create inventory movement
      description: Record inbound, outbound, or transfer movement. Validates stock availability for outbound/transfers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementRequest'
      responses:
        '201':
          description: Movement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryMovement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Concurrent update conflict or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /inventory/movements:
    get:
      tags: [Inventory]
      summary: List inventory movements
      parameters:
        - name: product_id
          in: query
          schema:
            type: string
            format: uuid
        - name: warehouse_id
          in: query
          schema:
            type: string
            format: uuid
        - name: movement_type
          in: query
          schema:
            type: string
            enum: [inbound, outbound, transfer]
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of movements
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryMovement'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'